fix(battacker-web): sync UI with battacker-extension
- Update App.tsx to match extension dashboard
- Add CyberGauge inline (remove separate component)
- Add initial-state, scan-hint, and interactive styles
